pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "kokoro-emotion-control"
version = "0.1.0"
description = "LLM-friendly emotion and prosody control for Kokoro TTS"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "soundfile",
  "torch",
  "PyYAML",
  "misaki",
]

[tool.setuptools]
packages = ["emotion_tts"]
